# src/main/resources/v1/data-catalog-openapi.yaml
openapi: 3.0.1
info:
  title: Data Catalog
  description: API for Data Catalog
  termsOfService: https://www.ericsson.com/en/legal
  contact:
    email: PDLEAMZEUG@pdl.internal.ericsson.com
  license:
    name: COPYRIGHT Ericsson "201"9
    url: https://www.ericsson.com/en/legal
  version: 1.0.0
servers:
- url: //localhost:8080/catalog
...
paths:
...
  /v1/data-space:
    get:
      tags:
      - /v1/data-space
      summary: get all dataspace details or filter by QueryParams
      operationId: queryDataspaceByParamsUsingGET
      parameters:
      - name: name
        in: query
        description: name of a dataspace
        schema:
          type: string
      responses:
        "200":
          description: get dataspace details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataSpaceDto'
        "400":
          description: Bad Request in request values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: dataspace details not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
    post:
      tags:
      - /v1/data-space
      summary: create dataspace details
      operationId: registerDataspaceUsingPOST
      requestBody:
        description: dataspace
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataSpaceDto'
        required: true
      responses:
        "201":
          description: created dataspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSpaceDto'
        "400":
          description: Bad Request in values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "409":
          description: dataspace details exists already
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
      x-codegen-request-body-name: dataspace model
  /v1/data-space/{id}:
    get:
      tags:
      - /v1/data-space
      summary: get dataspace details by id
      operationId: queryDataspaceByIdUsingGET
      parameters:
      - name: id
        in: path
        description: id uniquely identifies dataspace
        required: true
        schema:
          type: string
      responses:
        "200":
          description: get dataspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSpaceDto'
        "400":
          description: Bad Request in request values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: dataspace details not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
      deprecated: false
...
components:
  schemas:
...
    CatalogResponse:
      title: CatalogResponse
      type: object
      properties:
        message:
          type: string
        path:
          type: string
        timeStamp:
          type: string
          format: date-time
...
    DataSpaceDto:
      title: DataSpaceDto
      type: object
      properties:
        dataProviderTypeIds:
          type: array
          items:
            type: integer
            format: int32
        id:
          type: integer
          format: int32
        name:
          type: string
...